backend/requirements.txt
fastapi>=0.109.0
uvicorn[websockets]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.0
pydantic>=2.6.0
python-multipart>=0.0.9
elevenlabs>=1.0.0
cartesia>=1.0.0
openai>=1.0.0
pydub>=0.25.1
PyPDF2>=3.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.9
